# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CodeMind API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/upload/avatar": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "API endpoint for /api/upload/avatar",
        "description": "Tự động sinh tài liệu cho POST /api/upload/avatar. Hãy cập nhật mô tả chi tiết sau.",
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      },
      "delete": {
        "tags": [
          "Upload"
        ],
        "summary": "API endpoint for /api/upload/avatar",
        "description": "Tự động sinh tài liệu cho DELETE /api/upload/avatar. Hãy cập nhật mô tả chi tiết sau.",
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      }
    },
...
    "/api/user/code": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "API endpoint for /api/user/code",
        "description": "Tự động sinh tài liệu cho GET /api/user/code. Hãy cập nhật mô tả chi tiết sau.",
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "API endpoint for /api/user/code",
        "description": "Tự động sinh tài liệu cho POST /api/user/code. Hãy cập nhật mô tả chi tiết sau.",
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      }
    },
...
    "/api/users/[username]/saved": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "API endpoint for /api/users/[username]/saved",
        "description": "Tự động sinh tài liệu cho GET /api/users/[username]/saved. Hãy cập nhật mô tả chi tiết sau.",
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      }
    },
...
    "/api/users/profile": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "API endpoint for /api/users/profile",
        "description": "Tự động sinh tài liệu cho PUT /api/users/profile. Hãy cập nhật mô tả chi tiết sau.",
        "responses": {
          "200": {
            "description": "Thành công"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        }
      }
    },
...
  },
...
}